Fixed linting errors + major reformat
