+++ WatchParty/client/src/__test__/App.test.js
@@ -1,5 +1,3 @@
-import React from 'react';
-import { render } from '@testing-library/react';
 import App from '../App';
 
 test('renders learn react link', () => {
