--- WatchParty/client/src/components/LandingPage/LandingPage.js
@@ -1,17 +1,0 @@
-import React from 'react';
-
-import './LandingPage.css'
-
-class LandingPage extends React.Component{
-
-    render () {
-        return (
-            <div className="Landing">
-              <h1> LANDING PAGE</h1>
-            </div>
-        );
-    }
-
-}
-
-export default LandingPage;
