+++ WatchParty/client/src/components/LandingPage/LandingPage.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import './LandingPage.css'
+
+class LandingPage extends React.Component {
+  render() {
+    return (
+      <div className="Landing">
+        <h1> LANDING PAGE</h1>
+      </div>
+    );
+  };
+};
+
+export default LandingPage;
