+++ WatchParty/client/src/utils/history.js
@@ -1,4 +1,5 @@
 // src/utils/history.js
 
 import { createBrowserHistory } from "history";
+
 export default createBrowserHistory();
